+++ src/components/tables/amenitiesLists.tsx
@@ -1,26 +1,25 @@
 import Pagination from "../pagination";
-import { ReactNode, useCallback, useState } from "react";
+import { useCallback, useState } from "react";
 import Sort from "../filterAndSort/sort";
 import DeleteConfirmation from "../infoModal/delete-confirmation";
 import ModalTemplate from "../modal";
 import AddEditAmenities from "../amenities/create-amenities";
+import NoResult from "../noResult";
+import useGetAmenities from "../../services-hooks/useGetAmenities";
 
 export default function AmenitiesListsTable({
   header,
-  data,
   title,
 }: {
   header: string[];
-  data: {
-    id: number;
-    amentiesName: string;
-    icon: ReactNode;
-    text: string;
-  }[];
   title: string;
 }) {
   const [currentPage, setCurrentPage] = useState(1);
 
+  const { data, isLoading, isFailed, setIsFailed, retryFunction, pagination } =
+    useGetAmenities({ page: currentPage });
+
+  const [selectedId, setSelectedId] = useState("");
   const [openDelete, setOpenDelete] = useState(false);
   const [isDeleting, setIsDeleting] = useState(false);
   const [openEditAmenity, setOpenEditAmenity] = useState(false);
@@ -42,54 +41,61 @@ export default function AmenitiesListsTable({
           <h2 className="text-xl font-semibold">{title}</h2>
           <Sort id="room-options" label="Sort List" />{" "}
         </div>
-        <table className=" w-full overflow-x-auto">
-          <thead className="">
-            <tr className=" text-left bg-gray-200 text-gray-500 rounded-lg">
-              {header.map((head) => (
-                <th key={head}>{head}</th>
-              ))}
-            </tr>
-          </thead>
-          <tbody className="">
-            {data.map((request, index) => {
-              return (
-                <tr key={request?.id} className=" border-b">
-                  <td>{request?.amentiesName}</td>
-                  <td>{request?.icon}</td>
-                  <td>{request?.text}</td>
-                  <td className=" group relative">
-                    <span className=" p-2 text-lg">...</span>
-                    <span className="z-10 text-center group-hover:flex hidden w-52 bg-white text-sm absolute right-0 top-0 rounded-lg shadow-lg flex-col">
-                      <button
-                        type="button"
-                        onClick={() => setOpenEditAmenity(true)}
-                        className=" p-3 px-4 hover:bg-primary/10 transition-all rounded-lg"
-                      >
-                        Edit Amenity
-                      </button>
-                      <button
-                        type="button"
-                        onClick={() => setOpenDelete(true)}
-                        className="p-3 px-4 hover:bg-primary/10 transition-all rounded-lg"
-                      >
-                        Delete Amenity
-                      </button>
-                    </span>
-                  </td>
-                </tr>
-              );
-            })}
-          </tbody>
-        </table>
+        {data && data.length > 0 ? (
+          <table className=" w-full overflow-x-auto">
+            <thead className="">
+              <tr className=" text-left bg-gray-200 text-gray-500 rounded-lg">
+                {header.map((head) => (
+                  <th key={head}>{head}</th>
+                ))}
+              </tr>
+            </thead>
+            <tbody className="">
+              {data.map((request, index) => {
+                return (
+                  <tr key={request?.id} className=" border-b">
+                    <td>{request?.name}</td>
+                    <td>
+                      <img
+                        src={request?.image}
+                        alt={request?.name}
+                        title={request?.name}
+                        className=" w-10 h-auto"
+                      />
+                    </td>
+                    <td>{request?.description}</td>
+                    <td className=" group relative">
+                      <span className=" p-2 text-lg">...</span>
+                      <span className="z-10 text-center group-hover:flex hidden w-52 bg-white text-sm absolute right-0 top-0 rounded-lg shadow-lg flex-col">
+                        <button
+                          type="button"
+                          onClick={() => {
+                            setSelectedId(String(request?.id));
+                            setOpenEditAmenity(true);
+                          }}
+                          className=" p-3 px-4 hover:bg-primary/10 transition-all rounded-lg"
+                        >
+                          Edit Amenity
+                        </button>
+                        <button
+                          type="button"
+                          onClick={() => setOpenDelete(true)}
+                          className="p-3 px-4 hover:bg-primary/10 transition-all rounded-lg"
+                        >
+                          Delete Amenity
+                        </button>
+                      </span>
+                    </td>
+                  </tr>
+                );
+              })}
+            </tbody>
+          </table>
+        ) : (
+          <NoResult />
+        )}
         <Pagination
-          pagination={{
-            current_page: 1,
-            last_page: 2,
-            per_page: 20,
-            total: 24,
-            from: 1,
-            to: 1,
-          }}
+          pagination={pagination}
           setCurrentPage={setCurrentPage}
           isLoading={false}
           label="Amenities"
@@ -111,7 +117,7 @@ export default function AmenitiesListsTable({
         title="Edit Amenity"
         className=" max-w-md"
       >
-        <AddEditAmenities setOpen={setOpenEditAmenity} id="1" />
+        <AddEditAmenities setOpen={setOpenEditAmenity} id={selectedId} />
       </ModalTemplate>
     </>
   );