+++ src/services-hooks/useGetCustomerById.tsx
@@ -1,7 +1,6 @@
 import { useCallback, useEffect, useState } from "react";
 import useAxios from "../useHooks/useAxios";
 import { customersById } from "../types/apiData/customers";
-import tempAptData from "../assets/temp-api-mockup-data/customers.json";
 import { useAppDispatch } from "../stores/hooks";
 import {
   updateCustomerCompany,
@@ -24,95 +23,83 @@ export default function useGetCustomerById(id?: string) {
   const getCustomer = useCallback(async () => {
     setIsLoading(true);
     try {
-      // const response = await axios.get(`/show_apartment/${id}`);
-      // const responseData = response?.data?.data;
-      // setData(responseData);
-      const found = tempAptData.data.find((item) => item?.id === id);
-      if (found) {
-        const {
-          firstname,
-          lastname,
+      const response = await axios.get(`/admin/user/${id}`);
+      const { user } = response?.data?.data;
+      const {
+        id: resp_id,
+        first_name,
+        last_name,
+        email,
+        phone,
+        profile_photo,
+        dob,
+        gender,
+        identity_verification_document,
+        identity_verification_status,
+        has_set_password,
+        identity_verified,
+        email_verified_at,
+        created_at,
+        updated_at,
+        stripe_id,
+        pm_type,
+        pm_last_four,
+        trial_ends_at,
+        delete_reason,
+        referral_code,
+        referred_by,
+      } = user;
+      setData(user);
+      dispatch(updateCustomerInfoId({ id: id }));
+      dispatch(
+        updateCustomerDetails({
+          first_name,
+          last_name,
           email,
-          phoneNumber,
-          profileImg,
-          gender,
+          phone,
+          profileImg: { name: "", size: 1000, preview: profile_photo },
+          gender: gender || "Male",
           dob,
-          country,
-          state,
-          city,
-          address,
-          placeOfBirth,
-          idType,
-          idNumber,
-          idImage,
-          pinGenerated,
-          notes,
-          companyName,
-          VATid,
-          companyEmail,
-          companyId,
-          companyCountry,
-          companyState,
-          companyCity,
-          companyAddress,
-          isSalesChannel,
-          salesChannelName,
-          salesChannelCommision,
-          calculateCommissionOn,
-          applyCommissionOn,
-        } = found;
-        setData(found);
-        dispatch(updateCustomerInfoId({ id: id }));
-        dispatch(
-          updateCustomerDetails({
-            firstname,
-            lastname,
-            email,
-            phoneNumber,
-            profileImg: { name: "", size: 1000, preview: profileImg },
-            gender,
-            dob,
-            country,
-            state,
-            city,
-            address,
-          })
-        );
-        dispatch(
-          updateCustomerVerification({
-            placeOfBirth,
-            idType,
-            idNumber,
-            idImage,
-            pinGenerated,
-            notes,
-          })
-        );
-        dispatch(
-          updateCustomerCompany({
-            companyName,
-            VATid,
-            companyEmail,
-            companyId,
-            companyCountry,
-            companyState,
-            companyCity,
-            companyAddress,
-          })
-        );
-        dispatch(
-          updateCustomerSalesChannel({
-            isSalesChannel,
-            salesChannelName,
-            salesChannelCommision,
-            calculateCommissionOn,
-            applyCommissionOn,
-          })
-        );
-      }
+          country: "Nigeria",
+          state: "***",
+          city: "****",
+          address: "***",
+        })
+      );
+      dispatch(
+        updateCustomerVerification({
+          placeOfBirth: "***",
+          idType: "***",
+          idNumber: "****",
+          idImage: "****",
+          pinGenerated: "****",
+          notes: "***",
+        })
+      );
+      dispatch(
+        updateCustomerCompany({
+          companyName: "***",
+          VATid: "***",
+          companyEmail: "***",
+          companyId: "***",
+          companyCountry: "***",
+          companyState: "***",
+          companyCity: "***",
+          companyAddress: "***",
+        })
+      );
+      dispatch(
+        updateCustomerSalesChannel({
+          isSalesChannel: false,
+          salesChannelName: "***",
+          salesChannelCommision: "***",
+          calculateCommissionOn: "***",
+          applyCommissionOn: "***",
+        })
+      );
+
       setIsLoading(false);
     } catch (error) {
-      //   console.log({ error });
       setIsFailed(true);
     }
   }, [id]);